Crate fragments

Structs

Shell
A Shell is a layer on top of an overridable template.

Template
A string template with placeholders and conditional content.

Enums

ContentType
Container enum for template content

ExponentFormat
A flag that specifies whether to use exponential (scientific) notation.

SignificantDigits
The number of digits used for emitting the fractional part of a number, if any.

Token
Internal representation of template parts.

Traits

Generator
A trait for content generators.

InnerTemplate
A trait for overridable templates.

TemplateContent
A trait for types that can be inserted into templates