Enum fragments::ExponentFormatUnstable

pub enum ExponentFormat {
    ExpNone,
    ExpDec,
    ExpBin,
}

A flag that specifies whether to use exponential (scientific) notation.

Variants

ExpNone

Do not use exponential notation.

ExpDec

Use exponential notation with the exponent having a base of 10 and the exponent sign being e or E. For example, 1000 would be printed 1e3.

ExpBin

Use exponential notation with the exponent having a base of 2 and the exponent sign being p or P. For example, 8 would be printed 1p3.