Enum fragments::SignificantDigitsUnstable

pub enum SignificantDigits {
    DigAll,
    DigMax(usize),
    DigExact(usize),
}

The number of digits used for emitting the fractional part of a number, if any.

Variants

DigAll

All calculable digits will be printed.

Note that bignums or fractions may cause a surprisingly large number of digits to be printed.

DigMax

At most the given number of digits will be printed, truncating any trailing zeroes.

DigExact

Precisely the given number of digits will be printed.