Struct fragments::Shell

pub struct Shell<'r, 'c> {
    pub content: HashMap<String, Option<ContentType<'r>>>,
    pub generators: HashMap<String, Option<Box<Generator + 'r>>>,
    pub conditions: HashMap<String, bool>,
    // some fields omitted
}

A Shell is a layer on top of an overridable template.

Shells can be used to assign different values to a template and still keep the original intact.

Fields

content

Content for the placeholders

generators

Content generators

conditions

Conditional switches

Methods

impl<'r, 'c> Shell<'r, 'c>

fn new<T: InnerTemplate<'c>>(base: &'r T) -> Shell<'r, 'c>

Create a new Shell around base.

fn insert<T: TemplateContent<'c>>(&mut self, label: String, item: T)

Insert content.

fn unset(&mut self, label: String)

Unset content.

fn insert_formatted_float(&mut self, label: String, item: f64, precision: SignificantDigits, exponent: ExponentFormat)

Insert a formatted float.

fn insert_generator<T: Generator + Send>(&mut self, label: String, gen: T)

Insert a content generator.

fn unset_generator(&mut self, label: String)

Unset a content generator.

fn set(&mut self, label: String, value: bool)

Set a condition.

fn wrap<'a>(&'a self) -> Shell<'a, 'c>

Create an other Shell around this Shell.