Struct fragments::Template

pub struct Template<'c> {
    pub content: HashMap<String, ContentType<'c>>,
    pub generators: HashMap<String, Box<Generator + 'c>>,
    pub conditions: HashSet<String>,
    // some fields omitted
}

A string template with placeholders and conditional content.

Placeholders are written as [[:label]], where label becomes the name of the placeholder. The label is then used to insert content: my_template.insert("label", my_content);. The assigned content for a placeholder can be anything that implements Display. Even other templates may be inserted, which allows a more atomic structure.

Conditional segments are surrounded by [[?label]]...[[/]], where label becomes the name of the condition, and they are used to display content depending on whether its label exists in the conditions set. [[/]] marks the end of a block and may contain other characters after the /, which may be useful for labeling the end mark. Conditions can be made negative by writing [[?!label]]...[[/]], which makes the content visible if the label is missing from the conditions set. Conditional segments can also depend on whether a placeholder has an assigned value. Just write them like this: [[?:label]]...[[/]] or [[?!:label]]...[[/]].

Content can also be generated, using a generator token: [[+label arg1 arg2 ...]]. The label and the arguments are separated by one or more whitespaces. They can also be quoted to prevent special characters from being parsed: [[+"my label" arg1 "[[arg2]]"]]. The arguments will be passed to an instance of the Generator trait and the result will be inserted into the content.

Any character can be escaped by writing \ before it. It can be used like this: \[[[:label1]], [[:label2]]] which will result in [content1, content2], since the first [ will be ignored by the parser and added to the rest of the content.

Fields

content

Content for the placeholders

generators

Content generators

conditions

Conditional switches

Methods

impl<'c> Template<'c>

fn from_chars(b: Chars) -> Result<Template<'c>, String>

Create a new Template from a character iterator.

fn from_buffer<T: BufRead>(b: &mut T) -> Result<Template<'c>, String>

Create a new Template from a buffer.

fn insert<T: TemplateContent<'c>>(&mut self, label: String, item: T)

Insert content.

fn insert_formatted_float(&mut self, label: String, item: f64, precision: SignificantDigits, exponent: ExponentFormat)

Insert a formatted float.

fn insert_generator<T: Generator + Send + 'c>(&mut self, label: String, gen: T)

Insert a content generator.

fn set(&mut self, label: String, value: bool)

Set a condition.

fn wrap<'a>(&'a self) -> Shell<'a, 'c>

Create a Shell around this Template.

fn get_tokens(&self) -> &[Token]